refactor(headless): split Menu into per-file parts for tree-shaking
