+++ packages/headless/src/primitives/menu/index.ts
@@ -1,10 +1,12 @@
+export * as Menu from './parts';
+
 export type {
   MenuArrowProps,
   MenuItemProps,
   MenuPopupProps,
+  MenuPortalProps,
   MenuPositionerProps,
   MenuProps,
   MenuSeparatorProps,
   MenuTriggerProps,
-} from './menu';
-export { Menu } from './menu';
+} from './parts';
