+++ packages/headless/src/primitives/menu/parts.ts
@@ -1,0 +1,8 @@
+export { type MenuProps, MenuRoot as Root } from './menu-root';
+export { type MenuTriggerProps, MenuTrigger as Trigger } from './menu-trigger';
+export { type MenuPortalProps, MenuPortal as Portal } from './menu-portal';
+export { type MenuPositionerProps, MenuPositioner as Positioner } from './menu-positioner';
+export { type MenuPopupProps, MenuPopup as Popup } from './menu-popup';
+export { type MenuItemProps, MenuItem as Item } from './menu-item';
+export { type MenuSeparatorProps, MenuSeparator as Separator } from './menu-separator';
+export { type MenuArrowProps, MenuArrow as Arrow } from './menu-arrow';
